pyproject.toml
[project]
name = "discounted-stocks"
version = "0.1.0"
description = ""
authors = [
    {name = "Tejas K Jaiswal",email = "tejasjaiswal9711@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "yfinance (>=1.0,<2.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
packages = [{include = "discounted_stocks", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
